+++ packages/spot-consumer/src/streams/index.ts
@@ -1,1 +1,2 @@
 export * from './SpotMarketStream'
+export * from './SubaccountStream'